planner/ai/planner.py
from sklearn.linear_model import LinearRegression
import numpy as np

class LearningPlanGenerator:
    def __init__(self, user_data, course_data):
        self.user_data = user_data
        self.course_data = course_data

    def analyze_user_data(self):
        # Analyze user data to determine learning preferences and strengths
        # This is a placeholder for actual analysis logic
        return {
            'learning_style': 'visual',
            'strengths': ['math', 'science'],
            'weaknesses': ['literature']
        }

    def generate_plan(self):
        user_analysis = self.analyze_user_data()
        # Generate a personalized learning plan based on user analysis
        plan = {
            'recommended_courses': self.recommend_courses(user_analysis),
            'study_schedule': self.create_study_schedule(user_analysis)
        }
        return plan

    def recommend_courses(self, user_analysis):
        # Recommend courses based on user strengths and weaknesses
        recommended = []
        for course in self.course_data:
            if course['subject'] in user_analysis['strengths']:
                recommended.append(course['name'])
        return recommended

    def create_study_schedule(self, user_analysis):
        # Create a study schedule based on user preferences
        schedule = {}
        for course in self.recommend_courses(user_analysis):
            schedule[course] = '2 hours per week'
        return schedule

# Example usage
if __name__ == "__main__":
    user_data = {'id': 1, 'name': 'John Doe'}
    course_data = [
        {'name': 'Math 101', 'subject': 'math'},
        {'name': 'Science 101', 'subject': 'science'},
        {'name': 'Literature 101', 'subject': 'literature'}
    ]
    
    generator = LearningPlanGenerator(user_data, course_data)
    learning_plan = generator.generate_plan()
    print(learning_plan)